Small fixes (#2)
* refactor: Update package.json and remove unnecessary dependencies

Remove "hsl" and "@types/react-google-recaptcha" dependencies.

* refactor: Add "sharp" dependency for image processing

* refactor: Remove unnecessary SVG files

Remove unused SVG files for Vercel, Thirteen, and Next. These files are no longer needed and can be safely deleted.

* feat: Updated favicon

* refactor: Update Projects component styles

Optimize the Projects component styles by increasing the gap between elements and adjusting the text sizes for better readability. This improves the overall design and enhances the user experience.

* refactor: Optimize sticky scroll animation in Projects component

* refactor: Update Contact component styles

* refactor: Update header component styles

Adjust the font size of the menu links in the header component to improve readability. Also, update the menu open animation to include a smooth transition for the body overflow property.